requirements.txt
Django==4.2.7
django-ipware==6.0.3
fontawesomefree==6.5.1
ipdata==4.0.7
mysqlclient==2.2.0
mapcode==2.5.5.0
pycountry==23.12.11
python-dotenv==1.0.0
reverse_geocoder==1.5.1
